### 2024-01-22 #6258 @editorialbot
[PRE REVIEW]: BibDedupe: An Open-Source Python Library for Bibliographic Record Deduplication
**Submitting author:** @geritwagner (<a href="http://orcid.org/0000-0003-3926-7717">Gerit Wagner</a>)
**Repository:** https://github.com/CoLRev-Environment/bib-dedupe
**Branch with paper.md** (empty if default branch): paper
**Version:** 0.7.0
