+++ paper/paper.bib
@@ -48,7 +48,7 @@
   number={1},
   pages={79--95},
   year={2008},
-  doi={10.1177/109442810629663}
+  doi={10.1177/1094428106296638}
 }
 
 @article{KoumarelasEtAl2020,
