### 2024-01-22 #6258 @geritwagner
As reviewers, the following have expertise relevant to the submission: DrMattG (who focuses on systematic reviews and meta-analyis), khinsen (who reviewed pygetpapers), timClicks (who reviewed LISC), danielskatz (who served as an editor for LISC).

I just fixed the affiliation and DOI. Sorry for that. The project is released under an MIT license - not sure why the license file is not recognized.
