+++ paper/paper.bib
@@ -16,7 +16,8 @@
   volume={21},
   number={5},
   pages={10},
-  year={2020}
+  year={2020},
+  doi={10.17705/1jais.00630}
 }
 
 @software{DeBruin2019,
