### 2024-01-24 #6258 @geritwagner
Done. Thank you for the instructions - I did not know that the statistics are generated from the paper branch.
