+++ output/current_results.md
@@ -16,7 +16,7 @@
 | package    |   FP |   TP |   FN |   TN |   false_positive_rate |   specificity |   sensitivity |   precision |     f1 | runtime   |
 |:-----------|-----:|-----:|-----:|-----:|----------------------:|--------------:|--------------:|------------:|-------:|:----------|
 | asreview   |   14 |  425 |   11 | 1538 |                 0.009 |         0.991 |        0.9748 |      0.9681 | 0.9714 | 0:00:00   |
-| bib-dedupe |    0 |  408 |   28 | 1552 |                 0     |         1     |        0.9358 |      1      | 0.9668 | 0:00:17   |
+| bib-dedupe |    0 |  408 |   28 | 1552 |                 0     |         1     |        0.9358 |      1      | 0.9668 | 0:00:12   |
 
 ## digital_work (n=7159)
 
@@ -43,40 +43,41 @@
 
 | package    |   FP |   TP |   FN |   TN |   false_positive_rate |   specificity |   sensitivity |   precision |     f1 | runtime   |
 |:-----------|-----:|-----:|-----:|-----:|----------------------:|--------------:|--------------:|------------:|-------:|:----------|
-| bib-dedupe |    0 |  766 |    6 | 1084 |                0      |        1      |        0.9922 |      1      | 0.9961 | 0:00:08   |
+| bib-dedupe |    0 |  766 |    6 | 1084 |                0      |        1      |        0.9922 |      1      | 0.9961 | 0:00:09   |
 | asreview   |   12 |  772 |    0 | 1072 |                0.0111 |        0.9889 |        1      |      0.9847 | 0.9923 | 0:00:00   |
 
 ## srsr (n=53001)
 
 | package    |   FP |    TP |   FN |    TN |   false_positive_rate |   specificity |   sensitivity |   precision |     f1 | runtime   |
 |:-----------|-----:|------:|-----:|------:|----------------------:|--------------:|--------------:|------------:|-------:|:----------|
-| bib-dedupe |    0 | 16916 |   68 | 36005 |                0      |        1      |        0.996  |      1      | 0.998  | 0:12:14   |
+| bib-dedupe |    0 | 16916 |   68 | 36005 |                0      |        1      |        0.996  |      1      | 0.998  | 0:08:41   |
 | asreview   | 1781 | 16069 |  915 | 34224 |                0.0495 |        0.9505 |        0.9461 |      0.9002 | 0.9226 | 0:00:02   |
 
 ## diabetes (n=1845)
 
 | package    |   FP |   TP |   FN |   TN |   false_positive_rate |   specificity |   sensitivity |   precision |     f1 | runtime   |
 |:-----------|-----:|-----:|-----:|-----:|----------------------:|--------------:|--------------:|------------:|-------:|:----------|
-| bib-dedupe |    0 | 1261 |    0 |  584 |                0      |        1      |        1      |      1      | 1      | 0:00:11   |
+| bib-dedupe |    0 | 1261 |    0 |  584 |                0      |        1      |        1      |      1      | 1      | 0:00:09   |
 | asreview   |   24 | 1251 |   10 |  560 |                0.0411 |        0.9589 |        0.9921 |      0.9812 | 0.9866 | 0:00:00   |
 
 ## depression (n=79880)
 
 | package    |   FP |    TP |   FN |    TN |   false_positive_rate |   specificity |   sensitivity |   precision |     f1 | runtime   |
 |:-----------|-----:|------:|-----:|------:|----------------------:|--------------:|--------------:|------------:|-------:|:----------|
[…]
