+++ output/current_results.md
@@ -22,7 +22,7 @@
 
 | package    |   FP |   TP |   FN |   TN |   false_positive_rate |   specificity |   sensitivity |   precision |     f1 | runtime   |
 |:-----------|-----:|-----:|-----:|-----:|----------------------:|--------------:|--------------:|------------:|-------:|:----------|
-| bib-dedupe |    0 |  368 |    1 | 6790 |                 0     |         1     |        0.9973 |      1      | 0.9986 | 0:00:46   |
+| bib-dedupe |    0 |  368 |    1 | 6790 |                 0     |         1     |        0.9973 |      1      | 0.9986 | 0:00:45   |
 | asreview   |  217 |  343 |   26 | 6573 |                 0.032 |         0.968 |        0.9295 |      0.6125 | 0.7384 | 0:00:00   |
 
 ## haematology (n=1415)
@@ -43,14 +43,14 @@
 
 | package    |   FP |   TP |   FN |   TN |   false_positive_rate |   specificity |   sensitivity |   precision |     f1 | runtime   |
 |:-----------|-----:|-----:|-----:|-----:|----------------------:|--------------:|--------------:|------------:|-------:|:----------|
-| bib-dedupe |    0 |  766 |    6 | 1084 |                0      |        1      |        0.9922 |      1      | 0.9961 | 0:00:09   |
+| bib-dedupe |    0 |  766 |    6 | 1084 |                0      |        1      |        0.9922 |      1      | 0.9961 | 0:00:08   |
 | asreview   |   12 |  772 |    0 | 1072 |                0.0111 |        0.9889 |        1      |      0.9847 | 0.9923 | 0:00:00   |
 
 ## srsr (n=53001)
 
 | package    |   FP |    TP |   FN |    TN |   false_positive_rate |   specificity |   sensitivity |   precision |     f1 | runtime   |
 |:-----------|-----:|------:|-----:|------:|----------------------:|--------------:|--------------:|------------:|-------:|:----------|
-| bib-dedupe |    0 | 16916 |   68 | 36005 |                0      |        1      |        0.996  |      1      | 0.998  | 0:08:41   |
+| bib-dedupe |    0 | 16916 |   68 | 36005 |                0      |        1      |        0.996  |      1      | 0.998  | 0:08:32   |
 | asreview   | 1781 | 16069 |  915 | 34224 |                0.0495 |        0.9505 |        0.9461 |      0.9002 | 0.9226 | 0:00:02   |
 
 ## diabetes (n=1845)
@@ -64,14 +64,14 @@
 
 | package    |   FP |    TP |   FN |    TN |   false_positive_rate |   specificity |   sensitivity |   precision |     f1 | runtime   |
 |:-----------|-----:|------:|-----:|------:|----------------------:|--------------:|--------------:|------------:|-------:|:----------|
-| bib-dedupe |    0 | 10072 |   88 | 69712 |                0      |        1      |        0.9913 |      1      | 0.9957 | 0:11:59   |
+| bib-dedupe |    0 | 10072 |   88 | 69712 |                0      |        1      |        0.9913 |      1      | 0.9957 | 0:12:05   |
 | asreview   | 3377 |  6182 | 3978 | 66335 |                0.0484 |        0.9516 |        0.6085 |      0.6467 | 0.627  | 0:00:03   |
 
 ## neuroimaging (n=3438)
 
 | package    |   FP |   TP |   FN |   TN |   false_positive_rate |   specificity |   sensitivity |   precision |     f1 | runtime   |
 |:-----------|-----:|-----:|-----:|-----:|----------------------:|--------------:|--------------:|------------:|-------:|:----------|
[…]
