+++ output/current_results.md
@@ -43,7 +43,7 @@
 
 | package    |   FP |   TP |   FN |   TN |   false_positive_rate |   specificity |   sensitivity |   precision |     f1 | runtime   |
 |:-----------|-----:|-----:|-----:|-----:|----------------------:|--------------:|--------------:|------------:|-------:|:----------|
-| bib-dedupe |    0 |  766 |    6 | 1084 |                0      |        1      |        0.9922 |      1      | 0.9961 | 0:00:08   |
+| bib-dedupe |    0 |  766 |    6 | 1084 |                0      |        1      |        0.9922 |      1      | 0.9961 | 0:00:09   |
 | asreview   |   12 |  772 |    0 | 1072 |                0.0111 |        0.9889 |        1      |      0.9847 | 0.9923 | 0:00:00   |
 
 ## srsr (n=53001)
@@ -64,20 +64,20 @@
 
 | package    |   FP |    TP |   FN |    TN |   false_positive_rate |   specificity |   sensitivity |   precision |     f1 | runtime   |
 |:-----------|-----:|------:|-----:|------:|----------------------:|--------------:|--------------:|------------:|-------:|:----------|
-| bib-dedupe |    0 | 10072 |   88 | 69712 |                0      |        1      |        0.9913 |      1      | 0.9957 | 0:12:05   |
+| bib-dedupe |    0 | 10072 |   88 | 69712 |                0      |        1      |        0.9913 |      1      | 0.9957 | 0:12:06   |
 | asreview   | 3377 |  6182 | 3978 | 66335 |                0.0484 |        0.9516 |        0.6085 |      0.6467 | 0.627  | 0:00:03   |
 
 ## neuroimaging (n=3438)
 
 | package    |   FP |   TP |   FN |   TN |   false_positive_rate |   specificity |   sensitivity |   precision |     f1 | runtime   |
 |:-----------|-----:|-----:|-----:|-----:|----------------------:|--------------:|--------------:|------------:|-------:|:----------|
-| bib-dedupe |    0 | 1292 |    7 | 2139 |                0      |        1      |        0.9946 |      1      | 0.9973 | 0:00:20   |
+| bib-dedupe |    0 | 1292 |    7 | 2139 |                0      |        1      |        0.9946 |      1      | 0.9973 | 0:00:19   |
 | asreview   |    8 | 1266 |   33 | 2131 |                0.0037 |        0.9963 |        0.9746 |      0.9937 | 0.9841 | 0:00:00   |
 
 ## cardiac (n=8948)
 
 | package    |   FP |   TP |   FN |   TN |   false_positive_rate |   specificity |   sensitivity |   precision |     f1 | runtime   |
 |:-----------|-----:|-----:|-----:|-----:|----------------------:|--------------:|--------------:|------------:|-------:|:----------|
-| bib-dedupe |    0 | 3520 |   12 | 5415 |                0      |        1      |        0.9966 |      1      | 0.9983 | 0:01:00   |
+| bib-dedupe |    0 | 3520 |   12 | 5415 |                0      |        1      |        0.9966 |      1      | 0.9983 | 0:00:58   |
 | asreview   |  124 | 3176 |  356 | 5291 |                0.0229 |        0.9771 |        0.8992 |      0.9624 | 0.9297 | 0:00:00   |
 
