+++ output/current_results.md
@@ -22,14 +22,14 @@
 
 | package    |   FP |   TP |   FN |   TN |   false_positive_rate |   specificity |   sensitivity |   precision |     f1 | runtime   |
 |:-----------|-----:|-----:|-----:|-----:|----------------------:|--------------:|--------------:|------------:|-------:|:----------|
-| bib-dedupe |    0 |  368 |    1 | 6790 |                 0     |         1     |        0.9973 |      1      | 0.9986 | 0:00:46   |
+| bib-dedupe |    0 |  368 |    1 | 6790 |                 0     |         1     |        0.9973 |      1      | 0.9986 | 0:00:47   |
 | asreview   |  217 |  343 |   26 | 6573 |                 0.032 |         0.968 |        0.9295 |      0.6125 | 0.7384 | 0:00:00   |
 
 ## haematology (n=1415)
 
 | package    |   FP |   TP |   FN |   TN |   false_positive_rate |   specificity |   sensitivity |   precision |     f1 | runtime   |
 |:-----------|-----:|-----:|-----:|-----:|----------------------:|--------------:|--------------:|------------:|-------:|:----------|
-| bib-dedupe |    0 |  120 |   15 | 1278 |                 0     |         1     |        0.8889 |      1      | 0.9412 | 0:00:08   |
+| bib-dedupe |    0 |  120 |   15 | 1278 |                 0     |         1     |        0.8889 |      1      | 0.9412 | 0:00:09   |
 | asreview   |   46 |  119 |   16 | 1232 |                 0.036 |         0.964 |        0.8815 |      0.7212 | 0.7933 | 0:00:00   |
 
 ## stroke (n=1292)
@@ -50,7 +50,7 @@
 
 | package    |   FP |    TP |   FN |    TN |   false_positive_rate |   specificity |   sensitivity |   precision |     f1 | runtime   |
 |:-----------|-----:|------:|-----:|------:|----------------------:|--------------:|--------------:|------------:|-------:|:----------|
-| bib-dedupe |    0 | 16916 |   68 | 36005 |                0      |        1      |        0.996  |      1      | 0.998  | 0:08:42   |
+| bib-dedupe |    0 | 16916 |   68 | 36005 |                0      |        1      |        0.996  |      1      | 0.998  | 0:08:48   |
 | asreview   | 1781 | 16069 |  915 | 34224 |                0.0495 |        0.9505 |        0.9461 |      0.9002 | 0.9226 | 0:00:02   |
 
 ## diabetes (n=1845)
@@ -64,7 +64,7 @@
 
 | package    |   FP |    TP |   FN |    TN |   false_positive_rate |   specificity |   sensitivity |   precision |     f1 | runtime   |
 |:-----------|-----:|------:|-----:|------:|----------------------:|--------------:|--------------:|------------:|-------:|:----------|
-| bib-dedupe |    0 | 10072 |   88 | 69712 |                0      |        1      |        0.9913 |      1      | 0.9957 | 0:12:02   |
+| bib-dedupe |    0 | 10072 |   88 | 69712 |                0      |        1      |        0.9913 |      1      | 0.9957 | 0:12:07   |
 | asreview   | 3377 |  6182 | 3978 | 66335 |                0.0484 |        0.9516 |        0.6085 |      0.6467 | 0.627  | 0:00:03   |
 
 ## neuroimaging (n=3438)
@@ -78,6 +78,6 @@
 
 | package    |   FP |   TP |   FN |   TN |   false_positive_rate |   specificity |   sensitivity |   precision |     f1 | runtime   |
[…]
