### 2024-02-28 CoLRev-Environment/bib-dedupe#11 @DrMattG
Installation documentation 
https://github.com/openjournals/joss-reviews/issues/6318

For people who are non-"native" python users (like me!) it would be really nice to give more details about how to install the package. Maybe point to some useful resources that can get a user past the first hurdle... Many systematic reviewers want to try new tools but fall at the first hurdle of installation and understanding jargon around code.
