### 2024-02-28 CoLRev-Environment/bib-dedupe#11 @geritwagner
Thank you for raising this issue. I agree that the instructions should make the setup easier for users who are new to the Python ecosystem. Let me work on it and update the docs accordingly.
