### 2024-02-28 9ce16ea
add installation instructions to docs
