### 2024-02-28 CoLRev-Environment/bib-dedupe#11 @geritwagner
I have updated the docs with a separate installation page explaining the steps to setup Python on different platforms ([here](https://colrev-environment.github.io/bib-dedupe/installation.html)). I decided to merge/push to the main branch, which automatically updates the documentation. The setup steps currently take inspiration from the asreview project. Do you think the new page works?
