### 2024-02-29 CoLRev-Environment/bib-dedupe#13 @DrMattG
Example dataset?
https://github.com/openjournals/joss-reviews/issues/6318
Would it be possible to have an example dataset for users to use in the example workflow so they don't have to go and find a file in the correct format? I am not sure if you can add datasets to a python package like you can with R packages but there must be an elegant solution to that.
