### 2024-02-29 CoLRev-Environment/bib-dedupe#14 @DrMattG
AssertionError when I try and run the useage example
https://github.com/openjournals/joss-reviews/issues/6318#issuecomment-1970351682

No doubt this is down to my misunderstanding but I get this error when I try and run the deduplication:

`An error occurred while executing the following cell:
------------------
import pandas as pd
from bib_dedupe.bib_dedupe import merge

# Load your bibliographic dataset into a pandas DataFrame
records_df = pd.read_csv("data/duplicated.csv")

records_df

# Get the merged_df
merged_df = merge(records_df)
------------------

----- stdout -----
Loaded 1,895 records
Prep started at 2024-02-29 13:56:28
------------------

AssertionError                            Traceback (most recent call last)
Cell In[2], line 10
      7 records_df
      9 # Get the merged_df
---> 10 merged_df = merge(records_df)

File ~\AppData\Roaming\Python\Python312\site-packages\bib_dedupe\bib_dedupe.py:157, in merge(records_df, duplicate_id_sets, verbosity_level)
    154     verbose_print.verbosity_level = verbosity_level
    156 if not duplicate_id_sets:
--> 157     prep_df = prep(records_df)
    158     blocked_df = block(records_df=prep_df)
    159     matched_df = match(blocked_df)

File ~\AppData\Roaming\Python\Python312\site-packages\bib_dedupe\bib_dedupe.py:32, in prep(records_df, verbosity_level)
     30 if verbosity_level is not None:
     31     verbose_print.verbosity_level = verbosity_level
---> 32 return bib_dedupe.prep.prep(records_df)

File ~\AppData\Roaming\Python\Python312\site-packages\bib_dedupe\prep.py:200, in prep(records_df, cpu)
    197     verbose_print.print("No records to prepare")
    198     return {}
--> 200 records_df = __general_prep(records_df)
    201 cpu = determine_cpu_count(cpu, records_df.shape[0])
    202 df_split = np.array_split(records_df, cpu)

File ~\AppData\Roaming\Python\Python312\site-packages\bib_dedupe\prep.py:124, in __general_prep(records_df)
    121 if ID not in records_df.columns:
    122     records_df.loc[:, ID] = range(1, len(records_df) + 1)
--> 124 assert all(f in records_df.columns for f in REQUIRED_FIELDS)
    126 for column in records_df.columns:
    127     records_df[column] = records_df[column].replace(
    128         ["#NAME?", "UNKNOWN", ""], np.nan
    129     )

AssertionError:`

I created a csv in R by using the Gold standard diabetes dataset and artifically adding 50 duplicates

`
library(tidyverse)
# Duplicate some rows at random

Gold_Diabetes <- read_delim("data/Gold_Diabetes.csv",
                                 delim = ";", escape_double = FALSE, trim_ws = TRUE)

duplicates<-Gold_Diabetes |>
  sample_n(50)

duplicates<- duplicates |>
  bind_rows(Gold_Diabetes)

names(duplicates)<-c("ID","author", "year","journal","doi", "title", "pages","volume","number","abstract")

duplicates<-duplicates |>
  select(c(ID,author, year, journal,doi, title,pages, volume, number, abstract))

duplicates<-duplicates |>
  mutate(ID=1:n()) |>
  mutate("search_set"=NA)
duplicates |> view()

duplicates |> write.csv("data/duplicated.csv", row.names = FALSE)
`
