### 2024-02-29 b286ef3
required fields: add checks and print information (fixes #14)
