+++ bib_dedupe/block.py
@@ -210,6 +210,13 @@ def block(records_df: pd.DataFrame) -> pd.DataFrame:
     pd.DataFrame: The dataframe after blocking operation.
     """
 
+    assert (
+        "author_full" in records_df.columns
+    ), "Column 'author_full' not found in records_df (please pass the prepared df)"
+    assert (
+        "container_title" in records_df.columns
+    ), "Column 'container_title' not found in records_df (please pass the prepared df)"
+
     verbose_print.print(
         "Block started at " + datetime.now().strftime("%Y-%m-%d %H:%M:%S")
     )
