### 2024-02-29 CoLRev-Environment/bib-dedupe#14 @geritwagner
This was my fault. I assumed that the `ENTRYTYPE` field would always be available. Defaults are now used to cover cases without the `ENTRYTYPE` field. I also updated the error to provide better information. I updated the docs and implemented a few additional checks. Please feel free to reopen the issue if it persists.
