### 2024-02-29 884df8d
handle edge cases when merging with empty fields
