### 2024-02-29 fcd6bb1
add load_example_data for convenient access to data sets (closes #13)
