### 2024-02-29 d81d78e
add setuptools for pkg_resources
