+++ pyproject.toml
@@ -24,6 +24,7 @@ numpy = "^1.19.5"
 pandas = "^1.3.4"
 rapidfuzz = "^3.5.2"
 tqdm = "^4.66.1"
+setuptools = "^69.1.1"
 
 [tool.poetry.extras]
 with-data = ["pandas", "asreview", "notebook"]
