### 2024-02-29 CoLRev-Environment/bib-dedupe#13 @geritwagner
I just noticed that the data files are not distributed with the regular Python installation. I will post an update once I have a solution.
