### 2024-02-29 CoLRev-Environment/bib-dedupe#13 @geritwagner
Ok, now the datasets should be available. I just tested it successfully after installing the Package from GitHub.
