+++ output/current_results.md
@@ -2,12 +2,12 @@
 
 | package                    |   FP |    TP |   FN |     TN |   false_positive_rate |   specificity |   sensitivity |   precision |       f1 |
 |:---------------------------|-----:|------:|-----:|-------:|----------------------:|--------------:|--------------:|------------:|---------:|
-| bib-dedupe                 |    0 | 35035 |  227 | 125537 |           0           |      1        |      0.993562 |    1        | 0.996771 |
+| bib-dedupe                 |    0 | 35035 |  229 | 125541 |           0           |      1        |      0.993506 |    1        | 0.996742 |
 | Endnote (Hair et al. 2021) |   11 | 12466 | 3657 |  77977 |           0.000141047 |      0.999859 |      0.773181 |    0.999118 | 0.871748 |
 | Human (Hair et al. 2021)   |   54 | 14693 | 1430 |  77934 |           0.000692414 |      0.999308 |      0.911307 |    0.996338 | 0.951927 |
 | ASySD (Hair et al. 2021)   |   80 | 15664 |  459 |  77908 |           0.0010258   |      0.998974 |      0.971531 |    0.994919 | 0.983086 |
 | SRA-DM (Hair et al. 2021)  | 1735 | 12794 | 3329 |  76253 |           0.022247    |      0.977753 |      0.793525 |    0.880584 | 0.834791 |
-| asreview                   | 5615 | 29917 | 5345 | 119922 |           0.0447278   |      0.955272 |      0.84842  |    0.841973 | 0.845185 |
+| asreview                   | 5616 | 29918 | 5346 | 119925 |           0.0447344   |      0.955266 |      0.848401 |    0.841954 | 0.845165 |
 
 # Individual datasets
 
@@ -80,4 +80,11 @@
 |:-----------|-----:|-----:|-----:|-----:|----------------------:|--------------:|--------------:|------------:|-------:|:----------|
 | bib-dedupe |    0 | 3520 |   12 | 5415 |                0      |        1      |        0.9966 |      1      | 0.9983 | 0:01:01   |
 | asreview   |  124 | 3176 |  356 | 5291 |                0.0229 |        0.9771 |        0.8992 |      0.9624 | 0.9297 | 0:00:00   |
+
+## special_cases (n=6)
+
+| package    |   FP |   TP |   FN |   TN |   false_positive_rate |   specificity |   sensitivity |   precision |   f1 | runtime   |
+|:-----------|-----:|-----:|-----:|-----:|----------------------:|--------------:|--------------:|------------:|-----:|:----------|
+| asreview   |    1 |    1 |    1 |    3 |                  0.25 |          0.75 |           0.5 |         0.5 |  0.5 | 0:00:00   |
+| bib-dedupe |    0 |    0 |    2 |    4 |                  0    |          1    |           0   |         0   |  0   | 0:00:00   |
 
