### 2024-03-03 CoLRev-Environment/bib-dedupe#15 @DrMattG
Comments on manuscript
https://github.com/openjournals/joss-reviews/issues/6318

Line 68: "We define duplicates as potentially differing bibliographic representations..." - who is "we" in this sentence? (see also line 121; line 124; line 174; line 175, line 192)
Line 79: "A journal paper and a reprint..." - Why is this not considered a duplicate? This case strikes me as odd - it is the same study.
Tables - It would be good for the reader for the tables to have a caption and for them to be referenced in the text
Tables - can you explain (in the text or caption) what all the abbreviations mean.
