### 2024-03-03 CoLRev-Environment/bib-dedupe#16 @DrMattG
Supported bibliographic dataset formats
https://github.com/openjournals/joss-reviews/issues/6318

From the documentation it appears that only csv is currently supported. One of the things I like about ASySD is that there is a choice of upload formats (csv, bib, ris, endnote csv, endnote xml). Do you have plans to support more formats in the future?
