+++ paper/paper.md
@@ -77,7 +77,7 @@ Comparisons between platforms are not so relevant once users have decided for a 
 -->
 
 Open-source research software for duplicate identification is scarce, and to-date, peer-reviewed software is non-existent in this area.
-In the Python ecosystem, the only library we found is *ASReview Datatools*, provided by the team behind the *ASReview* screening tool [@VanDeSchootetAl2021].
+In the Python ecosystem, the only library I found is *ASReview Datatools*, provided by the team behind the *ASReview* screening tool [@VanDeSchootetAl2021].
 Our evaluations show that this library introduces a considerable number of false positives, and cannot be used for meta-analyses.
 R users or Python users willing to switch the ecosystem, may use ASySD [@Hair2023], a recently published R package with a Shiny web interface.
 The code of this package resembles BibDedupe, but it does not achieve zero-false-positives, uses a relatively small test dataset from medicine (n=1845) in the unit tests, and was not evaluated in the peer review process.
@@ -138,7 +138,7 @@ merged_df = merge(records_df, matched_df=matched_df)
 
 <!-- ## Duplicate definition -->
 
-We define duplicates as potentially differing bibliographic representations of the same real-world record [cf. @RathboneEtAl2015].
+I define duplicates as potentially differing bibliographic representations of the same real-world record [cf. @RathboneEtAl2015].
 This conceptual definition is operationalized as follows.
 The following are considered **duplicates**:
 
@@ -192,14 +192,14 @@ This is a common technique in deduplication where only records within the same b
 <!-- This significantly reduces the number of comparisons needed and improves the efficiency of the deduplication process. -->
 
 BibDedupe relies on a comprehensive set of blocking rules to avoid false negatives in this step.
-After the set of blocking rules is applied, we remove pairs not sharing a minimum number of words in the titles, effectively reducing the number of pairs by 50-95% without losing true pairs.
+After the set of blocking rules is applied, pairs not sharing a minimum number of words in the titles are removed, effectively reducing the number of pairs by 50-95% without losing true pairs.
 This leads to a more efficient matching step.
 
 ## Matching
 
 The matching function selects duplicates or potential duplicates from the list of blocked record pairs.
 Potential duplicates, also known as "maybe cases", are marked separately for manual verification.
-To achieve accurate and interpretable matching, we specified an array of human-readable conditions, which are based on pre-calculated and context-specific similarities between fields.
+To achieve accurate and interpretable matching, I specified an array of human-readable conditions, which are based on pre-calculated and context-specific similarities between fields.
 
 <!--
 Similarities (details):
@@ -207,9 +207,9 @@ Deduplication: continuous, additive, and context-agnostic similarity functions a
 In the specific context of bibliographic duplicates, it is essential to rationalize whether variation in meta-data can be explained by systems processing the same original record (including manuscript production systems, academic databases, or reference managers), or whether it can be explained by different publications.
 -->
 The conditions and similarity functions account for bibliographic errors commonly introduced between duplicates.
-We summarize the key design decisions of BibDedupe, which differ from other approaches (notably ASySD):
+I summarize the key design decisions of BibDedupe, which differ from other approaches (notably ASySD):
 
[…]
