### 2024-03-04 cfb61c6
update tables, captions, abbreviations
