+++ paper/paper.bib
@@ -177,3 +177,11 @@ year = {2022}
   year={2022},
   doi={10.1136/bmj-2022-072428},
 }
+
+@book{HigginsEtAl2023,
+  title={Cochrane Handbook for Systematic Reviews of Interventions version 6.4 (updated August 2023)},
+  author={Higgins, JPT and Thomas, J and Chandler, J and Cumpston, M and Li, T and Page, MJ and Welch, VA},
+  year={2023},
+  publisher={Cochrane},
+  url={www.training.cochrane.org/handbook}
+}
