### 2024-03-04 CoLRev-Environment/bib-dedupe#15 @geritwagner
Thank you for your comments on the manuscript. I addressed them as follows:

- replaced "we" with "I" (2fe8543bd251bb2f966953e7b7ee5c845615caef).
- included a paragraph explaining that different reports of the same study are not considered duplicates in this stage. As stated in the Cochrane handbook, it is good practice to link same studies after full-text retrieval, relying on information from the documents, external registers, manual assessment by the review team, and potentially involving the consultation of the authors of the original reports. (6a941f2d0bcdd3bf4c353eead8ea918f8736f97e).
- updated the tables with captions and abbreviations (cfb61c68f706adf8473b00009a6cb799b51046a5).

I hope my changes align with your expectations. I will leave the issue open for your responses or additional comments (if any).
