### 2024-03-04 CoLRev-Environment/bib-dedupe#16 @geritwagner
Thank you for raising this point. In short, I would prefer to keep the loading of file formats in a separate package and add code examples later. If you consider the support of CSV/BIB/RIS/XML important for the JOSS review, I can go ahead and implement it. Please let me know how to proceed.

**Detailed rationale** While the example shows how data can be loaded from CSV files, bib-dedupe generally supports Pandas dataframes as an input format. In principle, it should be possible to parse any file containing bibliographic metadata to this format. At the same time, databases do not implement the same standardized schema. For example, a RIS file produced by IEEEXplore stores data slightly different from a RIS file produced by Psycinfo (e.g., authors in the AU or A1 field). To properly reconcile different schemata, separate functionality would be needed, such as the [SearchSource packages](https://colrev.readthedocs.io/en/latest/manual/metadata_retrieval/search.html) currently developed in the CoLRev environment. Indeed, CoLRev is an example showing how bib-dedupe can be used programmatically (as a library), rather than requiring users to call it directly and handle different file formats. My plan is to support other formats (databases) based on CoLRev rather than implementing half-baked solutions in the bib-dedupe package. Once this is completed, I will add code examples to the bib-dedupe documentation, similar to the CSV case.
