### 2024-03-04 CoLRev-Environment/bib-dedupe#16 @DrMattG
Thanks for the detailed response - I do not think this is needed for the JOSS review - I guess it is quite a bit of extra work! Perhaps we can "document" (not the right word) this as an enhancement that will be implemented in the future.
