### 2024-03-04 CoLRev-Environment/bib-dedupe#16 @geritwagner
Superb. I have added a note to the documentation (61ee4b36aa2b40f73aa11648de667b7f73577fc7) and will update it once import functionality from CoLRev is ready.
