### 2024-03-08 CoLRev-Environment/bib-dedupe#18 @DrMattG
Cannot import name "load_example_data"
https://github.com/openjournals/joss-reviews/issues/6318

Probably me doing something wrong but I get this error: "ImportError: cannot import name 'load_example_data' from 'bib_dedupe.bib_dedupe' "

I have re-installed the package.
