### 2024-03-08 CoLRev-Environment/bib-dedupe#19 @DrMattG
Error with match
https://github.com/openjournals/joss-reviews/issues/6318

Again probably me misunderstanding python but I get this error when I run match - I was using this csv - "https://raw.githubusercontent.com/CoLRev-Environment/bib-dedupe/main/bib_dedupe/data/stroke/records_pre_merged.csv"

` >>> matched_df=match(blocked_df)
Sim started at 2024-03-08 09:33:59
C:\Users\matthew.grainger\AppData\Roaming\Python\Python312\site-packages\numpy\core\fromnumeric.py:59: FutureWarning: 'DataFrame.swapaxes' is deprecated and will be removed in a future version. Please use 'DataFrame.transpose' instead.
  return bound(*args, **kwds)
concurrent.futures.process._RemoteTraceback:
"""
Traceback (most recent call last):
  File "C:\Users\matthew.grainger\AppData\Roaming\Python\Python312\site-packages\pandas\core\indexes\base.py", line 3805, in get_loc
    return self._engine.get_loc(casted_key)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "index.pyx", line 167, in pandas._libs.index.IndexEngine.get_loc
  File "index.pyx", line 196, in pandas._libs.index.IndexEngine.get_loc
  File "pandas\\_libs\\hashtable_class_helper.pxi", line 7081, in pandas._libs.hashtable.PyObjectHashTable.get_item
  File "pandas\\_libs\\hashtable_class_helper.pxi", line 7089, in pandas._libs.hashtable.PyObjectHashTable.get_item
KeyError: 'author_full_1'

Traceback (most recent call last):
  File "C:\Program Files\Python312\Lib\concurrent\futures\process.py", line 263, in _process_worker
    r = call_item.fn(*call_item.args, **call_item.kwargs)
        ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "C:\Program Files\Python312\Lib\concurrent\futures\process.py", line 212, in _process_chunk
    return [fn(*args) for args in chunk]
            ^^^^^^^^^
  File "C:\Users\matthew.grainger\AppData\Roaming\Python\Python312\site-packages\bib_dedupe\sim.py", line 430, in process_df_split
    str(row["author_full_1"]),
        ~~~^^^^^^^^^^^^^^^^^
  File "C:\Users\matthew.grainger\AppData\Roaming\Python\Python312\site-packages\pandas\core\series.py", line 1112, in __getitem__
    return self._get_value(key)
           ^^^^^^^^^^^^^^^^^^^^
  File "C:\Users\matthew.grainger\AppData\Roaming\Python\Python312\site-packages\pandas\core\series.py", line 1228, in _get_value
    loc = self.index.get_loc(label)
          ^^^^^^^^^^^^^^^^^^^^^^^^^
  File "C:\Users\matthew.grainger\AppData\Roaming\Python\Python312\site-packages\pandas\core\indexes\base.py", line 3812, in get_loc
    raise KeyError(key) from err
KeyError: 'author_full_1'
"""

Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "C:\Users\matthew.grainger\AppData\Roaming\Python\Python312\site-packages\bib_dedupe\bib_dedupe.py", line 73, in match
    return bib_dedupe.match.match(pairs_df)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "C:\Users\matthew.grainger\AppData\Roaming\Python\Python312\site-packages\bib_dedupe\match.py", line 139, in match
    pairs = bib_dedupe.sim.calculate_similarities(pairs)
            ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "C:\Users\matthew.grainger\AppData\Roaming\Python\Python312\site-packages\bib_dedupe\sim.py", line 468, in calculate_similarities
    pairs_df = pd.concat(list(results))
                         ^^^^^^^^^^^^^
  File "C:\Program Files\Python312\Lib\concurrent\futures\process.py", line 642, in _chain_from_iterable_of_lists
    for element in iterable:
  File "C:\Program Files\Python312\Lib\concurrent\futures\_base.py", line 619, in result_iterator
    yield _result_or_cancel(fs.pop())
          ^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "C:\Program Files\Python312\Lib\concurrent\futures\_base.py", line 317, in _result_or_cancel
    return fut.result(timeout)
           ^^^^^^^^^^^^^^^^^^^
  File "C:\Program Files\Python312\Lib\concurrent\futures\_base.py", line 449, in result
    return self.__get_result()
           ^^^^^^^^^^^^^^^^^^^
  File "C:\Program Files\Python312\Lib\concurrent\futures\_base.py", line 401, in __get_result
    raise self._exception
KeyError: 'author_full_1'
>>>`
