### 2024-03-08 CoLRev-Environment/bib-dedupe#18 @geritwagner
Can you install the latest version from GitHub or the repository (using one the following commands)?

```
pip install git+https://github.com/CoLRev-Environment/bib-dedupe
pip install . # run in the bib-dedupe directory on main branch
```

With `pip install bib-dedupe` or `pip install .` on paper branch, an older version would be installed, potentially explaining the error. The `pip install bib-dedupe` (via PyPI) will install the latest changes once I create a new release. Does `load_example_data` with any of the two installation options above?
