### 2024-03-08 CoLRev-Environment/bib-dedupe#19 @geritwagner
Thank you for the erorr report. The current version should check for required columns earlier and raise a corresponding exception. Generally, if `author_full_1` is missing, it indicates that the `block()` was not applied at all or applied with errors. To reproduce the error, can you share the exact sequence of commands, the platform/Python version, and a printout (`print(blocked_df)`) of the dataframe before it is passed to the `machtch()`?
