+++ output/current_results.md
@@ -87,4 +87,3 @@
 |:-----------|-----:|-----:|-----:|-----:|----------------------:|--------------:|--------------:|------------:|-----:|:----------|
 | asreview   |    1 |    1 |    1 |    3 |                  0.25 |          0.75 |           0.5 |         0.5 |  0.5 | 0:00:00   |
 | bib-dedupe |    0 |    0 |    2 |    4 |                  0    |          1    |           0   |         0   |  0   | 0:00:00   |
-
