### 2024-03-18 CoLRev-Environment/bib-dedupe#18 @geritwagner
Wondering whether my suggestions are helpful. Please let me know if there is anything I can do to help with the next steps.
