### 2024-04-04 e9e0ac2
export_maybe: matche_df as keyword argument (for consistency)
