### 2024-04-04 CoLRev-Environment/bib-dedupe#19 @geritwagner
I found a Windows machine to test the issue. It turns out that the `prep()` was probably not applied, as suggested in the [usage example](https://colrev-environment.github.io/bib-dedupe/usage.html). This means I made a mistake in my previous comment when I speculated that `block()` was missing.

Entering the commands from the usage example with the `records_df = prep(records_df)` works on my machines.

I also updated the error message to provide clearer instructions. It now prints the following:

`Column 'author_full' not found in records_df (please run prep(records_df) and pass the prepared df)`.

Can you verify whether the example works with the `records_df = prep(records_df)` part?
