### 2024-04-04 #6318 @geritwagner
Hi @atrisovic,

I have a question related to [#18](https://github.com/CoLRev-Environment/bib-dedupe/issues/18). For this issue, it may be helpful to create a new release, making it possible to install the latest version through pip/PyPI. Would that be ok, or should I keep version 0.7.0, as indicated on top of the review ([here](https://github.com/openjournals/joss-reviews/issues/6318#issue-2116598313))?
