### 2024-04-10 #6318 @geritwagner
I had a look at other reviews. There are examples of projects where new releases were created during the review process (e.g., [pyGCodeDecode](https://github.com/FAST-LB/pyGCodeDecode)). Unless you prefer otherwise, I would go ahead and create a new release to make the installation and testing easier. Please let me know how to proceed.
