### 2024-04-12 60f4256
implement option without multiprocessing
