+++ docs/doc_requirements.txt
@@ -1,0 +1,3 @@
+sphinx-copybutton
+furo
+pandas
