+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "bib-dedupe"
-version = "0.7.0"
+version = "0.7.1"
 description = "Identify and merge duplicates in bibliographic records"
 authors = ["Gerit Wagner <gerit.wagner@uni-bamberg.de>"]
 license = "MIT"
