### 2024-04-12 0f933ad
block: prevent type/FutureWarning in non-parallel mode
