+++ CITATION.cff
@@ -5,6 +5,6 @@ authors:
   given-names: "Gerit"
   orcid: "https://orcid.org/0000-0003-3926-7717"
 title: "BibDedupe: An Open-Source Python Library for Bibliographic Record Deduplication"
-version: 0.7.1
+version: 0.7.2
 date-released: 2024-04-12
 url: "https://github.com/CoLRev-Environment/bib-dedupe"
