### 2024-04-16 13c2baf
retrieve example data on demand to avoid large package size
