+++ .pre-commit-config.yaml
@@ -15,9 +15,9 @@ repos:
     rev: v4.5.0
     hooks:
     -   id: trailing-whitespace
-        exclude: ^bib_dedupe/data/.*\.xml$
+        exclude: ^data/.*\.xml$
     -   id: end-of-file-fixer
-        exclude: ^tests/data/|^bib_dedupe/data/.*\.xml$
+        exclude: ^tests/data/|^data/.*\.xml$
     -   id: check-docstring-first
     -   id: check-json
     -   id: check-yaml
@@ -25,7 +25,7 @@ repos:
     -   id: check-toml
     -   id: debug-statements
     -   id: name-tests-test
-        exclude: ^bib_dedupe/data/parse.py
+        exclude: ^data/parse.py
 -   repo: https://github.com/psf/black
     rev: 23.10.1
     hooks:
@@ -52,7 +52,7 @@ repos:
     hooks:
     -   id: mypy
         args: [--disallow-untyped-defs, --disallow-incomplete-defs, --disallow-untyped-calls]
-        exclude: ^bib_dedupe/data/parse.py|^docs/source/conf.py
+        exclude: ^data/parse.py|^docs/source/conf.py
 - repo: https://github.com/astral-sh/ruff-pre-commit
   rev: v0.1.3
   hooks:
