+++ pyproject.toml
@@ -25,6 +25,7 @@ pandas = "^1.3.4"
 rapidfuzz = "^3.5.2"
 tqdm = "^4.66.1"
 setuptools = "^69.1.1"
+requests = "^2.31.0"
 
 [tool.poetry.extras]
 with-data = ["pandas", "asreview", "notebook"]
