### 2024-04-20 CoLRev-Environment/bib-dedupe#18 @geritwagner
As a quick update: I released new versions of BibDedupe to PyPI. This means `load_example_data` should now be available when running `pip install bib-dedupe`. Please let me know whether this solves the issue or whether it still persists.
