### 2024-04-21 2cc6ed4
fix special case in benchmark metrics
