+++ output/current_results.md
@@ -2,12 +2,12 @@
 
 | package                    |   FP |    TP |   FN |     TN |   false_positive_rate |   specificity |   sensitivity |   precision |       f1 |
 |:---------------------------|-----:|------:|-----:|-------:|----------------------:|--------------:|--------------:|------------:|---------:|
-| bib-dedupe                 |    0 | 35035 |  229 | 125541 |           0           |      1        |      0.993506 |    1        | 0.996742 |
+| bib-dedupe                 |    1 | 35035 |  229 | 125546 |           7.96514e-06 |      0.999992 |      0.993506 |    0.999971 | 0.996728 |
 | Endnote (Hair et al. 2021) |   11 | 12466 | 3657 |  77977 |           0.000141047 |      0.999859 |      0.773181 |    0.999118 | 0.871748 |
 | Human (Hair et al. 2021)   |   54 | 14693 | 1430 |  77934 |           0.000692414 |      0.999308 |      0.911307 |    0.996338 | 0.951927 |
 | ASySD (Hair et al. 2021)   |   80 | 15664 |  459 |  77908 |           0.0010258   |      0.998974 |      0.971531 |    0.994919 | 0.983086 |
 | SRA-DM (Hair et al. 2021)  | 1735 | 12794 | 3329 |  76253 |           0.022247    |      0.977753 |      0.793525 |    0.880584 | 0.834791 |
-| asreview                   | 5616 | 29918 | 5346 | 119925 |           0.0447344   |      0.955266 |      0.848401 |    0.841954 | 0.845165 |
+| asreview                   | 5618 | 29918 | 5346 | 119929 |           0.0447482   |      0.955252 |      0.848401 |    0.841907 | 0.845141 |
 
 # Individual datasets
 
@@ -22,14 +22,14 @@
 
 | package    |   FP |   TP |   FN |   TN |   false_positive_rate |   specificity |   sensitivity |   precision |     f1 | runtime   |
 |:-----------|-----:|-----:|-----:|-----:|----------------------:|--------------:|--------------:|------------:|-------:|:----------|
-| bib-dedupe |    0 |  368 |    1 | 6790 |                 0     |         1     |        0.9973 |      1      | 0.9986 | 0:00:47   |
+| bib-dedupe |    0 |  368 |    1 | 6790 |                 0     |         1     |        0.9973 |      1      | 0.9986 | 0:00:46   |
 | asreview   |  217 |  343 |   26 | 6573 |                 0.032 |         0.968 |        0.9295 |      0.6125 | 0.7384 | 0:00:00   |
 
 ## haematology (n=1415)
 
 | package    |   FP |   TP |   FN |   TN |   false_positive_rate |   specificity |   sensitivity |   precision |     f1 | runtime   |
 |:-----------|-----:|-----:|-----:|-----:|----------------------:|--------------:|--------------:|------------:|-------:|:----------|
-| bib-dedupe |    0 |  120 |   15 | 1278 |                 0     |         1     |        0.8889 |      1      | 0.9412 | 0:00:09   |
+| bib-dedupe |    0 |  120 |   15 | 1278 |                 0     |         1     |        0.8889 |      1      | 0.9412 | 0:00:08   |
 | asreview   |   46 |  119 |   16 | 1232 |                 0.036 |         0.964 |        0.8815 |      0.7212 | 0.7933 | 0:00:00   |
 
 ## stroke (n=1292)
@@ -50,7 +50,7 @@
 
 | package    |   FP |    TP |   FN |    TN |   false_positive_rate |   specificity |   sensitivity |   precision |     f1 | runtime   |
[…]
