+++ output/current_results.md
@@ -2,12 +2,12 @@
 
 | package                    |   FP |    TP |   FN |     TN |   false_positive_rate |   specificity |   sensitivity |   precision |       f1 |
 |:---------------------------|-----:|------:|-----:|-------:|----------------------:|--------------:|--------------:|------------:|---------:|
-| bib-dedupe                 |    1 | 35035 |  229 | 125546 |           7.96514e-06 |      0.999992 |      0.993506 |    0.999971 | 0.996728 |
+| bib-dedupe                 |    0 | 35036 |  229 | 125546 |           0           |      1        |      0.993506 |    1        | 0.996743 |
 | Endnote (Hair et al. 2021) |   11 | 12466 | 3657 |  77977 |           0.000141047 |      0.999859 |      0.773181 |    0.999118 | 0.871748 |
 | Human (Hair et al. 2021)   |   54 | 14693 | 1430 |  77934 |           0.000692414 |      0.999308 |      0.911307 |    0.996338 | 0.951927 |
 | ASySD (Hair et al. 2021)   |   80 | 15664 |  459 |  77908 |           0.0010258   |      0.998974 |      0.971531 |    0.994919 | 0.983086 |
 | SRA-DM (Hair et al. 2021)  | 1735 | 12794 | 3329 |  76253 |           0.022247    |      0.977753 |      0.793525 |    0.880584 | 0.834791 |
-| asreview                   | 5618 | 29918 | 5346 | 119929 |           0.0447482   |      0.955252 |      0.848401 |    0.841907 | 0.845141 |
+| asreview                   | 5617 | 29919 | 5346 | 119929 |           0.0447406   |      0.955259 |      0.848405 |    0.841935 | 0.845158 |
 
 # Individual datasets
 
@@ -50,7 +50,7 @@
 
 | package    |   FP |    TP |   FN |    TN |   false_positive_rate |   specificity |   sensitivity |   precision |     f1 | runtime   |
 |:-----------|-----:|------:|-----:|------:|----------------------:|--------------:|--------------:|------------:|-------:|:----------|
-| bib-dedupe |    0 | 16916 |   68 | 36005 |                0      |        1      |        0.996  |      1      | 0.998  | 0:08:38   |
+| bib-dedupe |    0 | 16916 |   68 | 36005 |                0      |        1      |        0.996  |      1      | 0.998  | 0:08:45   |
 | asreview   | 1781 | 16069 |  915 | 34224 |                0.0495 |        0.9505 |        0.9461 |      0.9002 | 0.9226 | 0:00:02   |
 
 ## diabetes (n=1845)
@@ -64,7 +64,7 @@
 
 | package    |   FP |    TP |   FN |    TN |   false_positive_rate |   specificity |   sensitivity |   precision |     f1 | runtime   |
 |:-----------|-----:|------:|-----:|------:|----------------------:|--------------:|--------------:|------------:|-------:|:----------|
-| bib-dedupe |    0 | 10072 |   88 | 69712 |                0      |        1      |        0.9913 |      1      | 0.9957 | 0:12:20   |
+| bib-dedupe |    0 | 10072 |   88 | 69712 |                0      |        1      |        0.9913 |      1      | 0.9957 | 0:12:18   |
 | asreview   | 3377 |  6182 | 3978 | 66335 |                0.0484 |        0.9516 |        0.6085 |      0.6467 | 0.627  | 0:00:03   |
 
 ## neuroimaging (n=3438)
@@ -78,13 +78,13 @@
 
 | package    |   FP |   TP |   FN |   TN |   false_positive_rate |   specificity |   sensitivity |   precision |     f1 | runtime   |
[…]
