+++ .github/workflows/evaluate.yml
@@ -25,6 +25,21 @@ jobs:
         run: |
           poetry install -E with-data
 
+      - name: Set up R
+        uses: r-lib/actions/setup-r@v2
+        with:
+          r-version: '4.2'
+
+      - name: Install ASySD with dependencies
+        run: |
+          sudo apt-get install -y libcurl4-openssl-dev libharfbuzz-dev libfribidi-dev
+          R -e 'install.packages(c("curl", "gert", "gh"))'
+          R -e 'install.packages(c("usethis", "pkgdown", "rcmdcheck", "rversions", "urlchecker"))'
+          R -e 'install.packages("devtools")'
+          R -e 'devtools::install_github("camaradesuk/ASySD")'
+        env:
+          GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
+
       - name: Run Evaluation
         run: |
           poetry run jupyter nbconvert --execute notebooks/evaluation.ipynb --to notebook --inplace
