### 2024-04-24 b9c2248
update asysd evaluation script
