### 2024-04-25 40b82a3
asysd: use commandArgs for benchmark-path
