+++ notebooks/asysd_evaluation.R
@@ -4,12 +4,11 @@
 library(ASySD)
 library(dplyr)
 
+filepath <- commandArgs(trailingOnly = TRUE)[1]
 # filepath <- "records_pre_merged.csv"
 
-filedir <- args[1]
-
 # load citations
-citation_data <- load_search(paste0(filedir, "records_pre_merged.csv"), method="csv") # endnote
+citation_data <- load_search(filepath, method="csv") # endnote
 
 # Set ID
 citation_data$record_id <- citation_data$ID
@@ -22,7 +21,7 @@ dedup_citations <- dedup_citations(citation_data)
 unique_citations <- dedup_citations$unique
 
 # Set ID
-citation_data$ID <- citation_data$record_id
+unique_citations$ID <- unique_citations$duplicate_id
 
 # Save unique_citations to CSV
 write.csv(unique_citations, file = "asysd_merged_df.csv", row.names = FALSE)
