+++ docs/evaluation.rst
@@ -1,13 +1,55 @@
 Evaluation
 ====================================
 
-Bib-dedupe is continuously evaluated against other Python libraries (currently the `asreview datatools <https://github.com/asreview/asreview-datatools>`_) for duplicate removal in bibliographic datasets.
-Complementary data from Hair et al. (2021) is added to the overview.
-The notebooks are available for the `evaluation <https://github.com/CoLRev-Environment/bib-dedupe/tree/main/notebooks/evaluation.ipynb>`_, and the datasets are available in the `data <data>`_ section.
-A summary of the evaluation is available in the `README.md <https://github.com/CoLRev-Environment/bib-dedupe/tree/main/notebooks/README.md>`_, aggregated summaries are exported to `current_results.md <https://github.com/CoLRev-Environment/bib-dedupe/tree/main/output/current_results.md>`_, and detailed results are exported to a `csv file <https://github.com/CoLRev-Environment/bib-dedupe/tree/main/output/evaluation.csv>`_.
+We run a continuous evaluation of tools for duplicate removal in bibliographic datasets.
+The `datasets <#datasets>`_ and `tools <#tools>`_ are summarized below.
+The `evaluation notebook <https://github.com/CoLRev-Environment/bib-dedupe/tree/main/notebooks/evaluation.ipynb>`_ is executed by a `GitHub workflow <https://github.com/CoLRev-Environment/bib-dedupe/actions/workflows/evaluate.yml>`_ on a weekly basis.
+Detailed results are exported to a `csv file <https://github.com/CoLRev-Environment/bib-dedupe/tree/main/output/evaluation.csv>`_, and aggregated summaries are exported to `current_results.md <https://github.com/CoLRev-Environment/bib-dedupe/tree/main/output/current_results.md>`_, 
 
 .. image:: _static/evaluation_total.png
    :alt: Evaluation
+
+Note: Complementary data from Hair et al. (2021) is added to the overview.
+
+.. _tools:
+
+Tools
+-------------------
+
+Tools are ranked according to false-positive rate. The lower the false-positive rate, the better the tool.
+
+.. list-table::
+   :header-rows: 1
+   :widths: 20 20 10 40
+
+   * - Tool
+     - Type
+     - Rank
+     - Comment
+   * - `BibDedupe <https://github.com/CoLRev-Environment/bib-dedupe>`_
+     - Python library
+     - 1
+     - 
+   * - `ASReview Data Dedup <https://github.com/asreview/asreview-datatools>`_
+     - Python library
+     - 2
+     - 
+   * - `ASySD <https://github.com/camaradesuk/ASySD>`_
+     - R package
+     - TBD
+      - 
+   * - Endnote
+     - Local software (proprietary)
+     - -
+     - No programmatic access, automated evaluation not possible
+   * - Covidence (proprietary)
+     - Web-based software (proprietary)
+     - -
+     - No programmatic access, automated evaluation not possible
+
+
+
+.. _datasets:
 
 Datasets
 -------------------
@@ -17,38 +59,38 @@ Datasets
    :widths: 40 20 40
 
    * - Dataset and reference
-     - Size
+     - Size (n)
      - Status
-   * - `digital_work [5] <https://github.com/CoLRev-Environment/bib-dedupe/tree/main/data/digital_work>`_
-     - n=7,159
-     - Included
    * - `cytology_screening [4] <https://github.com/CoLRev-Environment/bib-dedupe/tree/main/data/cytology_screening>`_
-     - n=1,856
+     - 1,856
      - Included (based on `OSF <https://osf.io/dyvnj/>`_)
[…]
