+++ docs/evaluation.rst
@@ -2,102 +2,108 @@ Evaluation
 ====================================
 
 We run a continuous evaluation of tools for duplicate removal in bibliographic datasets.
-The `datasets <#datasets>`_ and `tools <#tools>`_ are summarized below.
 The `evaluation notebook <https://github.com/CoLRev-Environment/bib-dedupe/tree/main/notebooks/evaluation.ipynb>`_ is executed by a `GitHub workflow <https://github.com/CoLRev-Environment/bib-dedupe/actions/workflows/evaluate.yml>`_ on a weekly basis.
-Detailed results are exported to a `csv file <https://github.com/CoLRev-Environment/bib-dedupe/tree/main/output/evaluation.csv>`_, and aggregated summaries are exported to `current_results.md <https://github.com/CoLRev-Environment/bib-dedupe/tree/main/output/current_results.md>`_, 
+Detailed results are exported to a `csv file <https://github.com/CoLRev-Environment/bib-dedupe/tree/main/output/evaluation.csv>`_, and aggregated summaries are exported to `current_results.md <https://github.com/CoLRev-Environment/bib-dedupe/tree/main/output/current_results.md>`_.
 
 .. image:: _static/evaluation_total.png
    :alt: Evaluation
 
 Note: Complementary data from Hair et al. (2021) is added to the overview.
 
-.. _tools:
-
-Tools
--------------------
-
 Tools are ranked according to false-positive rate. The lower the false-positive rate, the better the tool.
 
 .. list-table::
    :header-rows: 1
-   :widths: 20 20 10 40
+   :widths: 10 30 20 40
 
-   * - Tool
+   * - Rank
+     - Tool
      - Type
-     - Rank
      - Comment
-   * - `BibDedupe <https://github.com/CoLRev-Environment/bib-dedupe>`_
+   * - 1
+     - `BibDedupe <https://github.com/CoLRev-Environment/bib-dedupe>`_
      - Python library
-     - 1
-     - 
-   * - `ASReview Data Dedup <https://github.com/asreview/asreview-datatools>`_
-     - Python library
-     - 2
-     - 
-   * - `ASySD <https://github.com/camaradesuk/ASySD>`_
+     - NA
+   * - 2
+     - `ASySD <https://github.com/camaradesuk/ASySD>`_
      - R package
-     - TBD
-      - 
-   * - Endnote
+     - NA
+   * - 3
+     - `ASReview Data Dedup <https://github.com/asreview/asreview-datatools>`_
+     - Python library
+     - NA
+   * - NA
+     - Endnote
      - Local software (proprietary)
-     - -
-     - No programmatic access, automated evaluation not possible
-   * - Covidence (proprietary)
+     - No programmatic access (prevents automated evaluation)
+   * - NA
+     - Covidence
      - Web-based software (proprietary)
-     - -
-     - No programmatic access, automated evaluation not possible
+     - No programmatic access (prevents automated evaluation)
 
-
-
-.. _datasets:
 
 Datasets
 -------------------
 
 .. list-table::
    :header-rows: 1
-   :widths: 40 20 40
+   :widths: 20 20 20 40
 
-   * - Dataset and reference
+   * - Dataset
+     - Reference
      - Size (n)
      - Status
-   * - `cytology_screening [4] <https://github.com/CoLRev-Environment/bib-dedupe/tree/main/data/cytology_screening>`_
+   * - `cytology_screening <https://github.com/CoLRev-Environment/bib-dedupe/tree/main/data/cytology_screening>`_
+     - [4]
      - 1,856
      - Included (based on `OSF <https://osf.io/dyvnj/>`_)
[…]
