+++ notebooks/evaluation.py
@@ -86,256 +86,6 @@ def export_for_pytest(
     )
 
 
-def add_external_evaluations(dataset_df: pd.DataFrame) -> pd.DataFrame:
-    performance_data_depression = pd.DataFrame(
-        [
-            {
-                "package": "Human (Hair et al. 2021)",
-                "TP": 9390,
-                "TN": 69793,
-                "FN": 669,
-                "FP": 28,
-                "sensitivity": 0.933,
-                "specificity": 0.999,
-                "precision": 0.997,
-                "f1": 0.964,
-                "runtime": "Unknown",
-            },
-            {
-                "package": "Endnote (Hair et al. 2021)",
-                "TP": 7531,
-                "TN": 69816,
-                "FN": 2528,
-                "FP": 5,
-                "sensitivity": 0.749,
-                "specificity": 0.999,
-                "precision": 0.999,
-                "f1": 0.856,
-                "runtime": "<30 minutes",
-            },
-            {
-                "package": "SRA-DM (Hair et al. 2021)",
-                "TP": 9448,
-                "TN": 68473,
-                "FN": 611,
-                "FP": 1348,
-                "sensitivity": 0.939,
-                "specificity": 0.980,
-                "precision": 0.875,
-                "f1": 0.906,
-                "runtime": "~48 hours",
-            },
-            {
-                "package": "ASySD (Hair et al. 2021)",
-                "TP": 9624,
-                "TN": 69749,
-                "FN": 435,
-                "FP": 72,
-                "sensitivity": 0.957,
-                "specificity": 0.999,
-                "precision": 0.993,
-                "f1": 0.974,
-                "runtime": "<1 hour",
-            },
-        ]
-    )
-
-    performance_data_neuroimaging = pd.DataFrame(
-        [
-            {
-                "package": "Human (Hair et al. 2021)",
-                "TP": 1274,
-                "TN": 2139,
-                "FN": 19,
-                "FP": 6,
-                "sensitivity": 0.985,
-                "specificity": 0.997,
-                "precision": 0.996,
-                "f1": 0.990,
-                "runtime": "Unknown",
-            },
-            {
-                "package": "Endnote (Hair et al. 2021)",
-                "TP": 983,
-                "TN": 2142,
-                "FN": 310,
-                "FP": 3,
-                "sensitivity": 0.760,
-                "specificity": 0.999,
-                "precision": 0.997,
-                "f1": 0.863,
-                "runtime": "<5 minutes",
-            },
-            {
-                "package": "SRA-DM (Hair et al. 2021)",
-                "TP": 1050,
-                "TN": 2103,
-                "FN": 243,
-                "FP": 42,
-                "sensitivity": 0.812,
-                "specificity": 0.980,
-                "precision": 0.962,
-                "f1": 0.880,
-                "runtime": "<5 minutes",
-            },
-            {
-                "package": "ASySD (Hair et al. 2021)",
-                "TP": 1278,
-                "TN": 2141,
-                "FN": 15,
-                "FP": 4,
-                "sensitivity": 0.988,
-                "specificity": 0.998,
-                "precision": 0.997,
-                "f1": 0.993,
-                "runtime": "<5 minutes",
-            },
-        ]
-    )
-
-    performance_data_diabetes = pd.DataFrame(
-        [
-            {
-                "package": "Human (Hair et al. 2021)",
-                "TP": 893,
-                "TN": 581,
-                "FN": 368,
-                "FP": 3,
-                "sensitivity": 0.708,
-                "specificity": 0.995,
-                "precision": 0.997,
-                "f1": 0.828,
[…]
