+++ docs/evaluation.rst
@@ -1,10 +1,12 @@
 Evaluation
 ====================================
 
+.. image:: _static/evaluation_total.png
+   :alt: Evaluation
+
 We run a continuous evaluation of tools for duplicate removal in bibliographic datasets.
 The `evaluation notebook <https://github.com/CoLRev-Environment/bib-dedupe/tree/main/notebooks/evaluation.ipynb>`_ is executed by a `GitHub workflow <https://github.com/CoLRev-Environment/bib-dedupe/actions/workflows/evaluate.yml>`_ on a weekly basis.
 Detailed results are exported to a `csv file <https://github.com/CoLRev-Environment/bib-dedupe/tree/main/output/evaluation.csv>`_, and aggregated summaries are exported to `current_results.md <https://github.com/CoLRev-Environment/bib-dedupe/tree/main/output/current_results.md>`_.
-
 
 Tools are ranked according to false-positive rate. The lower the false-positive rate, the better the tool.
 
@@ -36,11 +38,6 @@ Tools are ranked according to false-positive rate. The lower the false-positive 
      - Covidence
      - Web-based software (proprietary)
      - No programmatic access (prevents automated evaluation)
-
-.. image:: _static/evaluation_total.png
-   :alt: Evaluation
-
-Note: Complementary data from Hair et al. (2021) is added to the overview.
 
 Datasets
 -------------------
