+++ .github/workflows/evaluate.yml
@@ -43,7 +43,7 @@ jobs:
       - name: Run Evaluation
         run: |
           poetry run jupyter nbconvert --execute notebooks/evaluation.ipynb --to notebook --inplace
-          poetry run notebooks/evaluation.py
+          poetry run python notebooks/evaluation.py
 
       - name: Commit and push changes
         run: |
