+++ output/current_results.md
@@ -1,14 +1,10 @@
 # Summary for datasets combined
 
-| package                    |   FP |    TP |   FN |     TN |   false_positive_rate |   specificity |   sensitivity |   precision |       f1 |
-|:---------------------------|-----:|------:|-----:|-------:|----------------------:|--------------:|--------------:|------------:|---------:|
-| bib-dedupe                 |    0 | 35036 |  229 | 125546 |           0           |      1        |      0.993506 |    1        | 0.996743 |
-| Endnote (Hair et al. 2021) |   11 | 12466 | 3657 |  77977 |           0.000141047 |      0.999859 |      0.773181 |    0.999118 | 0.871748 |
-| Human (Hair et al. 2021)   |   54 | 14693 | 1430 |  77934 |           0.000692414 |      0.999308 |      0.911307 |    0.996338 | 0.951927 |
-| asysd                      |   53 | 24464 |  641 |  55781 |           0.000949242 |      0.999051 |      0.974467 |    0.997838 | 0.986014 |
-| ASySD (Hair et al. 2021)   |   80 | 15664 |  459 |  77908 |           0.0010258   |      0.998974 |      0.971531 |    0.994919 | 0.983086 |
-| SRA-DM (Hair et al. 2021)  | 1735 | 12794 | 3329 |  76253 |           0.022247    |      0.977753 |      0.793525 |    0.880584 | 0.834791 |
-| asreview                   | 5617 | 29919 | 5346 | 119929 |           0.0447406   |      0.955259 |      0.848405 |    0.841935 | 0.845158 |
+| package    |   FP |    TP |   FN |     TN |   false_positive_rate |   specificity |   sensitivity |   precision |       f1 |
+|:-----------|-----:|------:|-----:|-------:|----------------------:|--------------:|--------------:|------------:|---------:|
+| bib-dedupe |    0 | 35036 |  229 | 125546 |           0           |      1        |      0.993506 |    1        | 0.996743 |
+| asysd      |   53 | 24464 |  641 |  55781 |           0.000949242 |      0.999051 |      0.974467 |    0.997838 | 0.986014 |
+| asreview   | 5617 | 29919 | 5346 | 119929 |           0.0447406   |      0.955259 |      0.848405 |    0.841935 | 0.845158 |
 
 # Individual datasets
 
@@ -24,7 +20,7 @@
 
 | package    |   FP |   TP |   FN |   TN |   false_positive_rate |   specificity |   sensitivity |   precision |     f1 | runtime   |
 |:-----------|-----:|-----:|-----:|-----:|----------------------:|--------------:|--------------:|------------:|-------:|:----------|
-| bib-dedupe |    0 |  368 |    1 | 6790 |                0      |        1      |        0.9973 |      1      | 0.9986 | 0:00:46   |
+| bib-dedupe |    0 |  368 |    1 | 6790 |                0      |        1      |        0.9973 |      1      | 0.9986 | 0:00:47   |
 | asysd      |   20 |  307 |   62 | 6770 |                0.0029 |        0.9971 |        0.832  |      0.9388 | 0.8822 | 0:00:15   |
 | asreview   |  217 |  343 |   26 | 6573 |                0.032  |        0.968  |        0.9295 |      0.6125 | 0.7384 | 0:00:00   |
 
@@ -32,7 +28,7 @@
 
 | package    |   FP |   TP |   FN |   TN |   false_positive_rate |   specificity |   sensitivity |   precision |     f1 | runtime   |
 |:-----------|-----:|-----:|-----:|-----:|----------------------:|--------------:|--------------:|------------:|-------:|:----------|
[…]
