+++ output/current_results.md
@@ -20,15 +20,15 @@
 
 | package    |   FP |   TP |   FN |   TN |   false_positive_rate |   specificity |   sensitivity |   precision |     f1 | runtime   |
 |:-----------|-----:|-----:|-----:|-----:|----------------------:|--------------:|--------------:|------------:|-------:|:----------|
-| bib-dedupe |    0 |  368 |    1 | 6790 |                0      |        1      |        0.9973 |      1      | 0.9986 | 0:00:47   |
-| asysd      |   20 |  307 |   62 | 6770 |                0.0029 |        0.9971 |        0.832  |      0.9388 | 0.8822 | 0:00:15   |
+| bib-dedupe |    0 |  368 |    1 | 6790 |                0      |        1      |        0.9973 |      1      | 0.9986 | 0:00:46   |
+| asysd      |   20 |  307 |   62 | 6770 |                0.0029 |        0.9971 |        0.832  |      0.9388 | 0.8822 | 0:00:14   |
 | asreview   |  217 |  343 |   26 | 6573 |                0.032  |        0.968  |        0.9295 |      0.6125 | 0.7384 | 0:00:00   |
 
 ## haematology (n=1415)
 
 | package    |   FP |   TP |   FN |   TN |   false_positive_rate |   specificity |   sensitivity |   precision |     f1 | runtime   |
 |:-----------|-----:|-----:|-----:|-----:|----------------------:|--------------:|--------------:|------------:|-------:|:----------|
-| bib-dedupe |    0 |  120 |   15 | 1278 |                 0     |         1     |        0.8889 |      1      | 0.9412 | 0:00:09   |
+| bib-dedupe |    0 |  120 |   15 | 1278 |                 0     |         1     |        0.8889 |      1      | 0.9412 | 0:00:08   |
 | asysd      |    0 |  109 |   26 | 1278 |                 0     |         1     |        0.8074 |      1      | 0.8934 | 0:00:04   |
 | asreview   |   46 |  119 |   16 | 1232 |                 0.036 |         0.964 |        0.8815 |      0.7212 | 0.7933 | 0:00:00   |
 
@@ -38,7 +38,7 @@
 |:-----------|-----:|-----:|-----:|-----:|----------------------:|--------------:|--------------:|------------:|-------:|:----------|
 | bib-dedupe |    0 |  312 |    2 |  978 |                0      |        1      |        0.9936 |      1      | 0.9968 | 0:00:07   |
 | asreview   |   12 |  314 |    0 |  966 |                0.0123 |        0.9877 |        1      |      0.9632 | 0.9812 | 0:00:00   |
-| asysd      |    0 |  302 |   12 |  978 |                0      |        1      |        0.9618 |      1      | 0.9805 | 0:00:04   |
+| asysd      |    0 |  302 |   12 |  978 |                0      |        1      |        0.9618 |      1      | 0.9805 | 0:00:03   |
 
 ## cytology_screening (n=1856)
 
@@ -52,8 +52,8 @@
 
 | package    |   FP |    TP |   FN |    TN |   false_positive_rate |   specificity |   sensitivity |   precision |     f1 | runtime   |
 |:-----------|-----:|------:|-----:|------:|----------------------:|--------------:|--------------:|------------:|-------:|:----------|
-| bib-dedupe |    0 | 16916 |   68 | 36005 |                0      |        1      |        0.996  |      1      | 0.998  | 0:08:53   |
-| asysd      |   25 | 16554 |  430 | 35980 |                0.0007 |        0.9993 |        0.9747 |      0.9985 | 0.9864 | 0:01:21   |
[…]
