+++ output/current_results.md
@@ -20,8 +20,8 @@
 
 | package    |   FP |   TP |   FN |   TN |   false_positive_rate |   specificity |   sensitivity |   precision |     f1 | runtime   |
 |:-----------|-----:|-----:|-----:|-----:|----------------------:|--------------:|--------------:|------------:|-------:|:----------|
-| bib-dedupe |    0 |  368 |    1 | 6790 |                0      |        1      |        0.9973 |      1      | 0.9986 | 0:00:46   |
-| asysd      |   20 |  307 |   62 | 6770 |                0.0029 |        0.9971 |        0.832  |      0.9388 | 0.8822 | 0:00:15   |
+| bib-dedupe |    0 |  368 |    1 | 6790 |                0      |        1      |        0.9973 |      1      | 0.9986 | 0:00:45   |
+| asysd      |   20 |  307 |   62 | 6770 |                0.0029 |        0.9971 |        0.832  |      0.9388 | 0.8822 | 0:00:14   |
 | asreview   |  217 |  343 |   26 | 6573 |                0.032  |        0.968  |        0.9295 |      0.6125 | 0.7384 | 0:00:00   |
 
 ## haematology (n=1415)
@@ -38,7 +38,7 @@
 |:-----------|-----:|-----:|-----:|-----:|----------------------:|--------------:|--------------:|------------:|-------:|:----------|
 | bib-dedupe |    0 |  312 |    2 |  978 |                0      |        1      |        0.9936 |      1      | 0.9968 | 0:00:07   |
 | asreview   |   12 |  314 |    0 |  966 |                0.0123 |        0.9877 |        1      |      0.9632 | 0.9812 | 0:00:00   |
-| asysd      |    0 |  302 |   12 |  978 |                0      |        1      |        0.9618 |      1      | 0.9805 | 0:00:04   |
+| asysd      |    0 |  302 |   12 |  978 |                0      |        1      |        0.9618 |      1      | 0.9805 | 0:00:03   |
 
 ## cytology_screening (n=1856)
 
@@ -46,14 +46,14 @@
 |:-----------|-----:|-----:|-----:|-----:|----------------------:|--------------:|--------------:|------------:|-------:|:----------|
 | bib-dedupe |    0 |  766 |    6 | 1084 |                0      |        1      |        0.9922 |      1      | 0.9961 | 0:00:09   |
 | asreview   |   12 |  772 |    0 | 1072 |                0.0111 |        0.9889 |        1      |      0.9847 | 0.9923 | 0:00:00   |
-| asysd      |    1 |  751 |   21 | 1083 |                0.0009 |        0.9991 |        0.9728 |      0.9987 | 0.9856 | 0:00:04   |
+| asysd      |    1 |  751 |   21 | 1083 |                0.0009 |        0.9991 |        0.9728 |      0.9987 | 0.9856 | 0:00:03   |
 
 ## srsr (n=53001)
 
 | package    |   FP |    TP |   FN |    TN |   false_positive_rate |   specificity |   sensitivity |   precision |     f1 | runtime   |
 |:-----------|-----:|------:|-----:|------:|----------------------:|--------------:|--------------:|------------:|-------:|:----------|
-| bib-dedupe |    0 | 16916 |   68 | 36005 |                0      |        1      |        0.996  |      1      | 0.998  | 0:08:47   |
-| asysd      |   25 | 16554 |  430 | 35980 |                0.0007 |        0.9993 |        0.9747 |      0.9985 | 0.9864 | 0:01:20   |
+| bib-dedupe |    0 | 16916 |   68 | 36005 |                0      |        1      |        0.996  |      1      | 0.998  | 0:08:41   |
[…]
