### 2024-04-29 eb59478
maybe_cases: prevent SettingWithCopyWarning
