+++ CITATION.cff
@@ -5,6 +5,6 @@ authors:
   given-names: "Gerit"
   orcid: "https://orcid.org/0000-0003-3926-7717"
 title: "BibDedupe: An Open-Source Python Library for Bibliographic Record Deduplication"
-version: 0.7.3
-date-released: 2024-04-16
+version: 0.7.4
+date-released: 2024-04-29
 url: "https://github.com/CoLRev-Environment/bib-dedupe"
