### 2024-05-01 CoLRev-Environment/bib-dedupe#25 @linuxscout
Error on example
Hi,
I tested the following code from documentation,
```python
from bib_dedupe.bib_dedupe import merge
from bib_dedupe.bib_dedupe import load_example_data

# Load example dataset
records_df = load_example_data("stroke")

# Get the merged_df
merged_df = merge(records_df)

# Save as csv
merged_df.to_csv("merged.csv", index=False)
```

I got:
```
Loaded 1,292 records
Prep started at 2024-05-01 11:11:58
Prep completed after: 0.80 seconds
Block started at 2024-05-01 11:11:59
Blocked      595 pairs with {'author_first', 'year'}
Blocked      465 pairs with {'title_short', 'volume'}
Blocked      506 pairs with {'container_title_short', 'author_first'}
Blocked      445 pairs with {'pages', 'title_short'}
Blocked      474 pairs with {'title_short', 'author_first'}
Blocked      450 pairs with {'container_title_short', 'title_short'}
Blocked    3,009 pairs with {'container_title_short', 'year', 'volume'}
Blocked    1,279 pairs with {'number', 'year', 'volume'}
Blocked      924 pairs with {'container_title_short', 'number', 'volume'}
Blocked      428 pairs with {'pages', 'container_title_short', 'volume'}
Blocked      480 pairs with {'title_short', 'year'}
Blocked      428 pairs with {'pages', 'container_title_short', 'year'}
Blocked      454 pairs with {'pages', 'year', 'volume'}
Blocked      431 pairs with {'pages', 'number', 'year'}
Blocked    3,568 pairs
Blocked pairs reduced to 583 pairs
Block completed after: 1.86 seconds
Sim started at 2024-05-01 11:12:01
Sim completed after: 0.13 seconds
Match started at 2024-05-01 11:12:01
Traceback (most recent call last):
  File "mergedb.py", line 8, in <module>
    merged_df = merge(records_df)
  File "/home/me/.local/lib/python3.8/site-packages/bib_dedupe/bib_dedupe.py", line 174, in merge
    matched_df = match(blocked_df)
  File "/home/me/.local/lib/python3.8/site-packages/bib_dedupe/bib_dedupe.py", line 84, in match
    return bib_dedupe.match.match(pairs_df, cpu=cpu)
  File "/home/me/.local/lib/python3.8/site-packages/bib_dedupe/match.py", line 163, in match
    true_pairs = __get_true_pairs(pairs)
  File "/home/me/.local/lib/python3.8/site-packages/bib_dedupe/match.py", line 87, in __get_true_pairs
    true_pairs = pairs.query("|".join(DUPLICATE_CONDITIONS))
  File "/home/me/.local/lib/python3.8/site-packages/pandas/core/frame.py", line 4113, in query
    res = self.eval(expr, **kwargs)
  File "/home/me/.local/lib/python3.8/site-packages/pandas/core/frame.py", line 4242, in eval
    return _eval(expr, inplace=inplace, **kwargs)
  File "/home/me/.local/lib/python3.8/site-packages/pandas/core/computation/eval.py", line 355, in eval
    ret = eng_inst.evaluate()
  File "/home/me/.local/lib/python3.8/site-packages/pandas/core/computation/engines.py", line 80, in evaluate
    res = self._evaluate()
  File "/home/me/.local/lib/python3.8/site-packages/pandas/core/computation/engines.py", line 121, in _evaluate
    return ne.evaluate(s, local_dict=scope)
  File "/home/me/.local/lib/python3.8/site-packages/numexpr/necompiler.py", line 836, in evaluate
    return compiled_ex(*arguments, **kwargs)
ValueError: too many inputs
```
Can you help me or fix the example.
