### 2024-05-01 53e80a5
add engine to query.join() in match to address #25
