### 2024-05-01 CoLRev-Environment/bib-dedupe#26 @linuxscout
Database formats
Hello,
I'm curious, can the library handle databases from other formats, such as BibTeX in LaTeX? Specifically, can users merge entries or remove duplicates?
Thanks.
