### 2024-05-01 #6318 @linuxscout
@geritwagner
Can the library handle databases from other formats, such as BibTeX in LaTeX?

https://github.com/CoLRev-Environment/bib-dedupe/issues/26
