### 2024-05-01 CoLRev-Environment/bib-dedupe#25 @geritwagner
Hello Taha,

thank you for helping with the review and for reporting the error. I just pushed a fix, using the python engine when passing the query conditions. Hopefully, this resolves the error. If not, could you share some details on the environment (OS, depdenency versions, especially numpy)?
