### 2024-05-01 CoLRev-Environment/bib-dedupe#26 @geritwagner
Good point. This is something we discussed in #16 . Let's continue the discussion in the original issue if you agree.
