### 2024-05-01 CoLRev-Environment/bib-dedupe#25 @linuxscout
I use python3.8 under linux,
pandas                            1.4.0
numpy                             1.20.1
