### 2024-05-01 #6318 @geritwagner
> @geritwagner please, fix this issue.
>
> > ```
> > Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
> >
> > OK DOIs
> >
> > - 10.1186/s12915-023-01686-z is OK
> > - 10.17705/1jais.00630 is OK
> > - 10.5281/zenodo.3559043 is OK
> > - 10.3163/1536-5050.104.3.014 is OK
> > - 10.1177/1094428106296638 is OK
> > - 10.1145/3377878 is OK
> > - 10.1186/2046-4053-4-6 is OK
> > - 10.1126/sciadv.abi8021 is OK
> > - 10.1186/1471-2288-9-10 is OK
> > - 10.3748/wjg.v23.i39.7198 is OK
> > - 10.1038/s42256-020-00287-7 is OK
> > - 10.1126/science.ade4585 is OK
> > - 10.1016/j.jsis.2021.101694 is OK
> > - 10.1136/bmj-2022-072428 is OK
> >
> > MISSING DOIs
> >
> > - No DOI given, and none found for title: dedupe
> > - No DOI given, and none found for title: Doing meta-analysis with R: A hands-on guide
> > - No DOI given, and none found for title: Improving our deduplication process
> > - No DOI given, and none found for title: NISO-RP-8-2008, Journal Article Versions (JAV): Re...
> > - No DOI given, and none found for title: Cochrane Handbook for Systematic Reviews of Interv...
> >
> > INVALID DOIs
> >
> > - None
> > ```

Thank you for checking the paper. I have just pushed an [update](https://github.com/CoLRev-Environment/bib-dedupe/commit/e53ab6ebf4d99a0c9aedf98a0f2f9bc4cf816942), adding the DOI for the NISO standard. For the others, I did not find DOIs (given that they are books or websites). I think it should be possible to cite such cases even though there is no DOI (in line with the examples in the [JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography)).
