### 2024-05-01 #6318 @linuxscout
it's ok.
> I think it should be possible to cite such cases even though there is no DOI
