### 2024-05-01 08820db
generally use engine=python in pandas.query()
pandas.query() throws a ValueError if numexpr is installed
see https://github.com/pandas-dev/pandas/issues/29027
