### 2024-05-01 CoLRev-Environment/bib-dedupe#25 @geritwagner
Thank you. I believe I was able to reproduce the issue on my machine. The error occurs if the numexpr package is installed (see [issue](https://github.com/pandas-dev/pandas/issues/29027)). To ensure that the behavior of `pandas.query()` is the same regardless of the numexpr package, I `engine="python"` to all `pandas.query()` calls. Please let me know if my changes resolve the issue for you.
