### 2024-05-01 CoLRev-Environment/bib-dedupe#18 @geritwagner
The import of `load_example_data` from `bib_dedupe.bib_dedupe` should work on the main branch and the current PyPI version (see #25). Please feel free to reopen the issue if you can confirm the ImportError in the current version on PyPI or the main branch.
