### 2024-05-01 #6318 @linuxscout
> @geritwagner َadd an issue, I ran the code example with some problems. [CoLRev-Environment/bib-dedupe#25](https://github.com/CoLRev-Environment/bib-dedupe/issues/25)

fixed
