+++ output/current_results.md
@@ -52,7 +52,7 @@
 
 | package    |   FP |    TP |   FN |    TN |   false_positive_rate |   specificity |   sensitivity |   precision |     f1 | runtime   |
 |:-----------|-----:|------:|-----:|------:|----------------------:|--------------:|--------------:|------------:|-------:|:----------|
-| bib-dedupe |    0 | 16916 |   68 | 36005 |                0      |        1      |        0.996  |      1      | 0.998  | 0:08:41   |
+| bib-dedupe |    0 | 16916 |   68 | 36005 |                0      |        1      |        0.996  |      1      | 0.998  | 0:08:39   |
 | asysd      |   25 | 16554 |  430 | 35980 |                0.0007 |        0.9993 |        0.9747 |      0.9985 | 0.9864 | 0:01:19   |
 | asreview   | 1781 | 16069 |  915 | 34224 |                0.0495 |        0.9505 |        0.9461 |      0.9002 | 0.9226 | 0:00:02   |
 
@@ -68,7 +68,7 @@
 
 | package    |   FP |    TP |   FN |    TN |   false_positive_rate |   specificity |   sensitivity |   precision |     f1 | runtime   |
 |:-----------|-----:|------:|-----:|------:|----------------------:|--------------:|--------------:|------------:|-------:|:----------|
-| bib-dedupe |    0 | 10072 |   88 | 69712 |                0      |        1      |        0.9913 |      1      | 0.9957 | 0:12:14   |
+| bib-dedupe |    0 | 10072 |   88 | 69712 |                0      |        1      |        0.9913 |      1      | 0.9957 | 0:12:13   |
 | asreview   | 3377 |  6182 | 3978 | 66335 |                0.0484 |        0.9516 |        0.6085 |      0.6467 | 0.627  | 0:00:03   |
 
 ## neuroimaging (n=3438)
@@ -92,6 +92,6 @@
 | package    |   FP |   TP |   FN |   TN |   false_positive_rate |   specificity |   sensitivity |   precision |     f1 | runtime   |
 |:-----------|-----:|-----:|-----:|-----:|----------------------:|--------------:|--------------:|------------:|-------:|:----------|
 | asreview   |    2 |    2 |    1 |    7 |                0.2222 |        0.7778 |        0.6667 |         0.5 | 0.5714 | 0:00:00   |
-| asysd      |    0 |    1 |    2 |    9 |                0      |        1      |        0.3333 |         1   | 0.5    | 0:00:02   |
+| asysd      |    0 |    1 |    2 |    9 |                0      |        1      |        0.3333 |         1   | 0.5    | 0:00:03   |
 | bib-dedupe |    0 |    1 |    2 |    9 |                0      |        1      |        0.3333 |         1   | 0.5    | 0:00:00   |
 
