### 2024-05-05 3fe9980
replace evaluation notebook with script for better runtime comparisons
