+++ pyproject.toml
@@ -17,7 +17,6 @@ python = "^3.8"
 pylint = {version = "3.0.1", optional = true }
 pytest = {version = "^7.2.1", optional = true }
 coverage = {extras = ["toml"], version = "^7.3.2", optional = true }
-notebook = {version = "^6.4.11", optional = true }
 asreview = {version = "^1.5", optional = true }
 number-parser = "^0.3.2"
 numpy = "^1.19.5"
@@ -28,7 +27,7 @@ setuptools = "^69.1.1"
 requests = "^2.31.0"
 
 [tool.poetry.extras]
-with-data = ["pandas", "asreview", "notebook"]
+with-data = ["pandas", "asreview"]
 dev = ["pylint", "pytest", "coverage", "pytest-skip-slow"]
 
 [tool.mypy]
