### 2024-05-06 #6318 @geritwagner
Thank you, @DrMattG! Your review and comments (especially those related to the setup and example datasets) were very helpful to make the first steps easier for users of BibDedupe.
