### 2024-05-16 d02aa3f
prep: check ID unique after ensuring that ID exists
