### 2024-05-20 #6318 @geritwagner
Hi @atrisovic, thank you - that is great news! I have just released version ``0.8.0``, which contains all of the latest changes. I have also fixed the metadata structure in the paper so that the ORCID and affiliation are displayed correctly. The Zenodo DOI is ``10.5281/zenodo.11223590``. I have verified that the author information, including ORCID, and license are correct and identical.
